+++ extensions/validator/src/commands/view_command/view_proposals/contract/download_mode/hash_contract/mod.rs
