+++ extensions/validator/src/commands/view_command/view_proposals/contract/mod.rs
