+++ extensions/validator/src/commands/view_command/view_validators/epoch/mod.rs
@@ -2,7 +2,6 @@ use dialoguer::Input;
 
 #[derive(Debug, Clone, clap::Clap)]
 pub enum CliSendTo {
-    /// Specify an account
     Account(CliSender),
 }
