+++ extensions/validator/src/commands/view_command/view_validators/mod.rs
@@ -1,3 +1,3 @@
 mod block_id;
 pub mod operation_mode;
-mod sender;
+mod epoch;
